chore: move rightclick tests into own file
